backend/requirements.txt
Django==3.2
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-filter==22.1
django-mptt==0.14.0
django-resized==1.0.2
djangorestframework==3.14.0
drf-spectacular==0.25.1
Pillow==9.4.0
psycopg2==2.8.6
PyJWT==2.6.0
gunicorn==20.1.0